fix name generator type issue

+++ src/db/name_generator.rs
@@ -110,7 +110,7 @@ impl IdEncoder for Harsh {
 
 pg_query!( GetNextId =>
     in = ;
-    out = id: i32;
+    out = id: i64;
     sql = r#"
         SELECT nextval('user_id_counter')
     "#
